전체: Debug 빌드 경고를 0 으로 만들고 그 과정에서 드러난 버그를 고친다
Ecs / CvSim / ScSim / HostSim / HostSim_DB 다섯 솔루션을 Debug|Win32 로
전체 재빌드하면 경고가 137 건 나왔다. 전부 정리해 0 오류 0 경고로 만든다.

경고가 실제 버그였던 것
- HostSim PasswordDlg : 유니코드 빌드인데 char 버퍼에 narrow sprintf 로
  넓은 문자열을 넣고 다시 LPCWSTR 로 캐스팅해, 바꾼 비밀번호가 깨진 채
  ini 에 저장됐다. 버퍼를 없애고 CString 을 그대로 넘긴다. (C4477)
- HostSim LogDlg : 로그 저장/인쇄가 같은 이유로 각 항목의 첫 글자만
  찍혔다. 저장은 파일에 바이트를 그대로 쓰므로 CT2A 로 ANSI 변환하고,
  인쇄 제목은 TCHAR 버퍼와 _stprintf_s 로 바꿔 캐스팅을 없앤다. (C4477)
- EcsLib MinButton::SetFontName : strcpy(CT2A(대상), ...) 이라 임시 변환
  버퍼에 써버려 폰트 이름이 반영되지 않았다. _tcscpy_s 로 고친다.
  (현재 호출처는 없어 동작 변화는 없다)

서식 인자 불일치 (C4477)
- Ecs / HostSim_DB msjexhnd : 덤프 경로 CString 을 %s 에 그대로 넘기던
  것에 LPCTSTR 캐스트를 넣는다

형 변환 (C4244)
- CvSim EcsView / LogicValidationDlg : GetTotalSeconds() 는 double 이라
  int 대입에 캐스트를 명시한다
- ScSim EcsView : 레지스터가 WORD 라 하위 바이트만 쓴다는 뜻을 캐스트로 밝힌다

BOOL 과 bool 혼용 비교 (C4805)
- HostSim_DB 5개 파일 9곳의 == true 를 == TRUE 로 바꾼다 (값은 동일)

죽은 선언 / 무시되는 지정자
- C4101 : DciLib DciControl, CvSim DeviceMap/LogicValidationDlg,
  ScSim Sc, HostSim_DB Equipment 의 안 쓰는 지역 변수를 뺀다
- C4091 : EcsLib FooButton/MinButton 의 static enum 에서 무시되는 static 제거
- C4008 : XmlLib dllmain 의 DllMain 은 DLL 진입점이라 static 이 무시된다

폐기 API (C4996)
- CRT : strcpy/wcscpy/strcat/wcscat/sprintf 를 _s 버전으로 바꾼다
  (대상이 고정 배열이라 템플릿 오버로드가 크기를 잡는다).
  StartupTip 은 &Tips[p].Text[0] 가 포인터로 붕괴해 크기를 잃던 것도 같이 고친다
- Enable3dControls : MFC 7 이후 기본 동작이라 호출을 지운다
- winsock : inet_addr/gethostbyname/inet_ntoa 는 대체 API 가 오류 처리
  동작을 바꾸므로 EcsLib 에 _WINSOCK_DEPRECATED_NO_WARNINGS 를 정의한다
- BCMenu(2002년 벤더 코드) : GetVersion 은 Win95~XP 판별용이라
  VersionHelpers 로 바꾸면 열거값이 달라진다. 함수 범위로만 경고를 끈다

빌드 설정
- D9035 : 더 이상 지원되지 않는 /Gm(MinimalRebuild) 을 끈다
  (ScSim, HostSim, BuilderLib)
- MSB8028 : Bin/Debug/CvSim 에 남아 있던 8월 7~8일자 ScSim 산출물 잔재 제거
  (현재 ScSim 의 IntDir 은 Bin/Debug/ScSim 이다)

Release 구성 일부 정정 (아직 링크되지는 않는다)
- 문자셋이 Release 만 MultiByte 로 남아 코드가 컴파일조차 안 되던 것을
  Debug 와 같이 Unicode 로 맞춘다
- ScSim/HostSim 의 Release 인클루드 경로가 Lib 폴더 재구성 이전 값
  (../EcsLib) 이던 것을 실제 위치로 고친다

+++ CvSim/EcsView.cpp
@@ -1064,7 +1064,7 @@ BOOL CEcsView::InvokeLogic(CCv* pCv, CTrackInfo* pTrack, CLogicArray* pLogicArra
 				int a = 0;
 
 			int nSec = _ttoi(strKeyWord);
-			int nTotalSec = tElapseTime.GetTotalSeconds();
+			int nTotalSec = (int)tElapseTime.GetTotalSeconds();
 			BOOL bIsDestSelf = pDoc->IsDestination(strKeyWord, nGetData, pTrack->m_nNumber, 1);
 			BOOL bIsScDest = pDoc->IsDestination(strKeyWord, nGetData, 0, 2);
 			BOOL bIsStation = pDoc->IsStationKind(pTrack, strKeyWord);
@@ -3683,7 +3683,7 @@ void CEcsView::MoveNextTrackForKindNormal_1(CCv* pCv, CTrackInfo* pTrack, CStati
 		
 		//CTimeSpan tElapseTime = COleDateTime::GetCurrentTime() - pTrack->m_tTime;
 		COleDateTimeSpan tElapseTime = COleDateTime::GetCurrentTime() - pTrack->m_tTime;
-		int nElapse = tElapseTime.GetTotalSeconds();
+		int nElapse = (int)tElapseTime.GetTotalSeconds();
 		if (nElapse > 2)
 		{
 			pTrack->m_bStartFlag = FALSE;
